Fixed errors.
log4j.properties will be part of jar. Log levels can be changed from there.
Jar and bat should be in the same directory.
Logs will be captured in applog.txt in the same directory.
DB info are present on the application.properties file.
Separate sql files are for created for different scenario.